refactor: move useControllableState to _internal for shared use
